refactor([issue-757]): extract subcomponents from ScheduleTab
Split the 1088-LOC ScheduleTab into a thin orchestrator plus a
schedule/ subdirectory of focused subcomponents: IntervalBadge,
PromptEditor, PipelineStageConfig, GlobalConfigControls,
AppOverrideRow, PerAppOverrideList, AppTaskTypeRow, AppTaskTypeSection,
and a shared scheduleConstants module. ScheduleTab is now 182 LOC.
Behavior unchanged; default export path preserved.
